zpu: rework file layout and naming
Moving forward, all applications will be prefixed with zpu_ while
any included compilation units will not have that. Primarily for
ease of organization and differentiation of file purpose.

Introduces muxbus.c/h. Include files intended to handle MUXBUS
interaction directly from the ZPU.

Introduces ts8820.h. Constants used on the TS-8820 specifically
for use with MUXBUS.

Modifies ts_zpu.h to provide some additional macros to ease
application interaction.

With the introduction/modification of the above, applications that
previously directly had this code in the sources have all of that
removed and are now set up to compile and include the above as
needed.
