+++ src/zpu/strings.h
@@ -1,10 +1,10 @@
 /* SPDX-License-Identifier: BSD-2-Clause */
 
-#ifndef __ZPU_STRINGS_H__
-#define __ZPU_STRINGS_H__
+#ifndef __STRINGS_H__
+#define __STRINGS_H__
 
 int printf(const char *format, ...);
 int sprintf(char *pInto, const char *format, ...);
 
-#endif //__ZPU_STRINGS_H__
+#endif //__STRINGS_H__
