+++ src/zpu/zpu_muxbus.c
@@ -3,40 +3,9 @@
 #include <stdarg.h>
 #include <string.h>
 
-#include "zpu_fifo.h"
+#include "muxbus.h"
+#include "fifo.h"
 #include "ts_zpu.h"
-
-/* NOTE: In this current implementation, only 16-but MUXBUS access are supported.
- * This is meant specifically for the TS-8820 which has 16-bit registers only.
- */
-
-#define READ		1
-#define WRITE		0
-
-/* AD bits are 31:27 of REG1, and 10:0 of REG2 */
-#define ALEn_bit	26
-#define DIR_bit		25
-#define CSn_bit		24
-#define BHEn_bit	23
-#define WAIT_bit	22
-
-/* MUXBUS timing definitions have been in number of clocks of the main BB clk.
- * In the case of the TS-8820, these are 12.5 MHz. MUXBUS guidelines can be
- * found here: https://wiki.embeddedTS.com/wiki/Generic_MUXBUS
- *
- * The ZPU in the TS-4100 operates at 63 MHz, and has a free running counter
- * that delay times are based on. This means that every 5.04 63 MHz clocks is
- * 1 12.5 MHz clock. Round up to 6 to be safe. Additionally, all of the delay
- * clks are + 1 in the final application
- *
- * The numbers below are based on a 0xF0FF value in the standard MUXBUS config
- * register.
- */
-#define TP_ALE		(0x07 + 1)	* 6
-#define TH_ADR		(0x21 + 1)	* 6
-#define TSU_DAT		(0x03 + 1)	* 6
-#define TP_CS		(0x03 + 1)	* 6
-#define TH_DAT		(0x03 + 1)	* 6
 
 /* State machine defines for operation loop */
 #define	GET_CMD		0
@@ -46,190 +15,6 @@
 #define	GET_DATL	4
 #define	RET_WRITE	5
 #define	RET_READ	6
-
-/* MUXBUS packet construction
- *
- * NOTE: in this current implementation, only 16-bit MUXBUS accesses are
- * supported.
- *
- * The request packet is 3 or 5 bytes. The first byte being a configuration
- * packet, followed by the 16-bit MUXBUS address, and in the case of a write,
- * 16-bit data to write to the address.
- *
- * 3 or 5 bytes
- * MSB - 0
- *   bit 0: 1 = MB Read, 0 = MB Write
- *   bit 1: 1 = 16bit, 0 = 8bit
- *   bit 7-2: Number of reads to stream + 1
- *     A value from 0 to 63. Will cause <value>+1 muxbus transactions
- *     which will end up with <value>+1 * 2 bytes written to the TX buffer.
- *   This value only has meaning for reads, it is unused for MUXBUS writes.
- * MSB - 1:2
- *   MUXBUS address
- * MSB - 3:4
- *   MUXBUS data (only during MB Write)
- *
- * Response 2 bytes
- *   2:1 Read data
- *
- * After a successful MUXBUS read or write, the ZPU will assert an IRQ to the CPU.
- * In the case of a read, the 2 bytes are put in the FIFO before an IRQ is raised.
- * In the case of a write, no data is put in the ZPU TX FIFO, but an IRQ is still
- * asserted. This allows the CPU to wait until it can be assured the MUXBUS write
- * was completed. This is safe since the CPU side FIFO read would clear the IRQ,
- * and simply find that there was no new data in the buffer.
- */
-
-void initmuxbusio(void)
-{
-	/* Enable 12.5 MHz clock on DIO_03 pin via FPGA reg 'd87 */
-	O_REG2 |= (1 << 23);
-
-	/* Set all AD pins to input for now */
-	OE_REG1 &= ~(0xF8000000);
-	OE_REG2 &= ~(0x000007FF);
-
-	/* ALE#, CS#, DIR, and BHE# are always outputs
-	 * Start them all off as high, aka deasserted. */
-	O_REG1 |= ((1 << ALEn_bit) | (1 << CSn_bit) | (1 << DIR_bit) |
-	  (1 << BHEn_bit));
-	OE_REG1 |= ((1 << ALEn_bit) | (1 << CSn_bit) | (1 << DIR_bit) |
-	  (1 << BHEn_bit));
-}
-
-void set_alen(unsigned long val)
-{
-	if (val) {
-		O_REG1 |= (1 << ALEn_bit);
-	} else {
-		O_REG1 &= ~(1 << ALEn_bit);
-	}
-}
-
-void set_dir(unsigned long val)
-{
-	if (val) {
-		O_REG1 |= (1 << DIR_bit);
-	} else {
-		O_REG1 &= ~(1 << DIR_bit);
-	}
-}
-
-void set_csn(unsigned long val)
-{
-	if (val) {
-		O_REG1 |= (1 << CSn_bit);
-	} else {
-		O_REG1 &= ~(1 << CSn_bit);
-	}
-}
-
-/* The TS-8820 actually ignores BHE# and only accepts 16bit accesses. This line
- * is unused in this application, but here for completeness
- */
-void set_bhen(unsigned long val)
-{
-	if (val) {
-		O_REG1 |= (1 << BHEn_bit);
-	} else {
-		O_REG1 &= ~(1 << BHEn_bit);
-	}
-}
-
-unsigned long get_wait(void)
-{
-	return ((O_REG1 & (1 << WAIT_bit)) >> 21);
-}
-
-void set_ad(unsigned short dat)
-{
-	O_REG1 &= ~(0xF8000000);
-	O_REG1 |= ((unsigned long)(dat << 27 ) & 0xF8000000);
-
-	O_REG2 &= ~(0x000007FF);
-	O_REG2 |= ((unsigned long)(dat >> 5) & 0x000007FF);
-}
-
-void set_ad_oe(unsigned long dir)
-{
-	if (dir) {
-		OE_REG1 |= (0xF8000000);
-		OE_REG2 |= (0x000007FF);
-	} else {
-		OE_REG1 &= ~(0xF8000000);
-		OE_REG2 &= ~(0x000007FF);
-	}
-}
-
-unsigned short get_ad(void)
-{
-	unsigned short dat = 0;
-
-	dat = ((I_REG2 & 0x000007FF) << 5);
-	dat |= ((I_REG1 & 0xF8000000) >> 27);
-
-	return dat;
-}
-
-/* This works so long as the maximum delay count is not more than half the span
- * of the free running counter. The counter in this is 32 bits wide, so limit
- * the max delay to 16 bits. In practice, this MUXBUS application will come
- * nowhere near that max.
- */
-void delay_clks(unsigned short cnt)
-{
-	unsigned long end_time;
-
-	end_time = TIMER_REG + cnt;
-	while ((signed long)(end_time - TIMER_REG) > 0);
-}
-
-/* The following functions are unused in this application as we normally
- * interleave reading bytes from the ZPU FIFO and doing MUXBUS accesses. Doing so
- * allows for a slightly smaller memory footprint
- * The functions are left in place for completeness, but are not compiled during
- * normal use. */
-#if 0
-void muxbus_write_16(unsigned short adr, unsigned short dat)
-{
-	set_dir(WRITE);
-	set_ad(adr);
-	set_ad_oe(1);
-	set_alen(0);
-	delay_clks(TP_ALE);
-	set_alen(1);
-	delay_clks(TH_ADR);
-	set_ad(dat);
-	delay_clks(TSU_DAT);
-	set_csn(0);
-	delay_clks(TP_CS);
-	set_csn(1);
-	delay_clks(TH_DAT);
-}
-
-unsigned short muxbus_read_16(unsigned short adr)
-{
-	unsigned short dat;
-
-	set_dir(READ);
-	set_ad(adr);
-	set_ad_oe(1);
-	set_alen(0);
-	delay_clks(TP_ALE);
-	set_alen(1);
-	delay_clks(TH_ADR);
-	set_ad_oe(0);
-	delay_clks(TSU_DAT);
-	set_csn(0);
-	delay_clks(TP_CS);
-	dat = get_ad();
-	set_csn(1);
-	delay_clks(TH_DAT);
-
-	return dat;
-}
-#endif
-
 
 /* ZPU MUXBUS application.
  *
